--- .idea/.gitignore
@@ -1,3 +1,0 @@
-
-# Default ignored files
-/workspace.xml
